Add reference for type an minor corrections
For both `is` and `is not` I have added reference that `type` references the Terraform resource type as this is very useful to know and wasn't clear from the documentation. I've tidied up the descriptions around those too (not requiring "e.g" in the description when it's a separate column). Whilst doing this I also fixed two spelling mistakes and capitalised "Terraform" when referring to the program and not the command line.