Add parse options git branch and directory
Users may want to reference a single directory from a git repository
or a ref that is not `master`.

This now follows the Terraform Modules in package sub-directories
format for referncing a ref in git.
https://www.terraform.io/docs/modules/sources.html#modules-in-package-sub-directories

Note that if the string passed in via `-f` does not end in `.git`
it must now contain both `.git//` and `?ref=`

It does not currently do anything with these inputs -- it still behaves as before.